third_party/blink/web_tests/platform/generic/http/tests/security/cross-origin-shared-worker-allowed-expected.txt

This iframe should not return any errors:

This iframe should not return any errors:

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Connected successfully

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Connected successfully
